backend/promise-fruit/pyproject.toml
[project]
name = "promise-fruit"
version = "0.1.0"
description = ""
authors = [
    {name = "JaeUk",email = "siwall0105@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiofiles (>=24.1.0,<25.0.0)"
]

[tool.poetry]
packages = [{include = "promise_fruit", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
